fix: Cleanup forking model
